move EntityTypes to ~/types
