+++ frontend/src/scenes/insights/ActionFilter/ActionFilterDropdown.tsx
@@ -2,7 +2,7 @@ import React, { RefObject } from 'react'
 import { BuiltLogic, useActions, useValues } from 'kea'
 import { ActionType } from '~/types'
 import { EventUsageType } from '~/types'
-import { EntityTypes } from '../../trends/trendsLogic'
+import { EntityTypes } from '~/types'
 import { actionsModel } from '~/models/actionsModel'
 import { FireOutlined, InfoCircleOutlined, AimOutlined, ContainerOutlined } from '@ant-design/icons'
 import { Tooltip } from 'antd'
